# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hæstore API",
    "version": "0.2.0",
    "description": "The catalogue is admin-defined: categories, attributes and variants are data, not code. Endpoints that return filters or attribute sets describe whatever an administrator has configured, so a storefront built against this document needs no hardcoded knowledge of any particular product kind."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Product": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProductCard"
          },
          {
            "type": "object",
            "properties": {
              "description": {
                "type": "string"
              },
              "category": {
                "type": "string"
              },
              "categoryAncestors": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "status": {
                "type": "string",
                "enum": [
                  "draft",
                  "active",
                  "archived"
                ]
              },
              "attributes": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ProductAttribute"
                }
              },
              "variantAxes": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "variants": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Variant"
                }
              },
              "defaultVariantId": {
                "type": "string"
              },
              "axes": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "unit": {
                      "type": "string"
                    },
                    "options": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "value": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "swatchHex": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "value",
                          "label"
                        ]
                      }
                    }
                  },
                  "required": [
                    "key",
                    "label",
                    "options"
                  ]
                },
                "description": "Each axis in variantAxes with its label and its values’ labels and swatches, so a picker never has to show the raw slug a variant stores."
              }
            },
            "required": [
              "category",
              "categoryAncestors",
              "status",
              "attributes",
              "variantAxes",
              "variants",
              "axes"
            ]
          }
        ]
      },
...
      "FacetValue": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "swatchHex": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "description": "Computed as if this group’s own filter were absent, so a value the shopper has not ticked shows what ticking it as well would return. A declared value that currently matches nothing is 0 rather than missing."
          },
          "selected": {
            "type": "boolean"
          }
        },
        "required": [
          "value",
          "label",
          "count",
          "selected"
        ]
      },
      "Facet": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "roast"
          },
          "label": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "select",
              "multiselect",
              "text",
              "number",
              "boolean",
              "color",
              "dimension"
            ]
          },
          "filterUi": {
            "type": "string",
            "enum": [
              "checkbox",
              "swatch",
              "range",
              "toggle",
              "select"
            ]
          },
          "unit": {
            "type": "string"
          },
          "range": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "min": {
                "type": "number"
              },
              "max": {
                "type": "number"
              }
            },
            "required": [
              "min",
              "max"
            ],
            "description": "Bounds for a range control, of what is available rather than of what is selected — so a slider can always be widened again. Null for value facets."
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetValue"
            }
          }
        },
        "required": [
          "key",
          "label",
          "type",
          "filterUi",
          "range",
          "values"
        ]
      },
      "ListingCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "priceRange": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "min": {
                "type": "integer"
              },
              "max": {
                "type": "integer"
              },
              "currency": {
                "type": "string"
              }
            },
            "required": [
              "min",
              "max",
              "currency"
            ]
          },
          "inStock": {
            "type": "boolean"
          },
          "image": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "publicId": {
                "type": "string"
              },
              "alt": {
                "type": "string"
              },
              "width": {
                "type": "integer"
              },
              "height": {
                "type": "integer"
              },
              "blurDataUrl": {
                "type": "string"
              }
            },
            "required": [
              "publicId",
              "alt"
            ]
          },
          "ratingAverage": {
            "type": "number"
          },
          "ratingCount": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "priceRange",
          "inStock",
          "image",
          "ratingAverage",
          "ratingCount"
        ]
      },
...
      "AdminCategory": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Category"
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "hidden"
                ]
              },
              "validationMode": {
                "type": "string",
                "enum": [
                  "lenient",
                  "strict"
                ]
              },
              "attributeBindings": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "defId": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "required": {
                      "type": "boolean"
                    },
                    "order": {
                      "type": "integer"
                    },
                    "group": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "defId",
                    "key",
                    "required",
                    "order"
                  ]
                }
              },
              "suppressedKeys": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "children": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "description": "AdminCategory nodes, recursively. Described loosely because the document format cannot express the recursion without a reference cycle."
              }
            },
            "required": [
              "status",
              "validationMode",
              "attributeBindings",
              "suppressedKeys",
              "children"
            ]
          }
        ]
      },
...
    },
...
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hae_sid",
        "description": "Opaque 256-bit session id, looked up in Redis. There are no JWTs anywhere in this system. The __Host- prefix forbids a Domain attribute, so a compromised subdomain cannot set a session cookie for the API."
      }
    }
  },
  "paths": {
...
    "/api/catalog/products": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "The storefront listing: search, filters, sorting and facet counts.",
        "description": "Served by Meilisearch, with a MongoDB fallback behind the same URL. `page.degraded` says which answered: when it is true, `facets` is null and attribute filters were not applied — each one comes back in `ignoredFilters` with a reason.\n\n**Any parameter not listed here is treated as an attribute filter**, matched against the category's own AttributeDefinitions. That is how a filter an admin defined this morning works without a deploy: `?roast=medium,dark&weight_g=250-1000`. Values within one attribute are OR’d; different attributes are AND’d. An unrecognised key or value is reported in `ignoredFilters` rather than rejected, so a bookmark outlives the attribute it names.\n\nPage size defaults to 24 and is capped at 60; depth is capped at 1000 documents.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Full-text query."
            },
            "required": false,
            "description": "Full-text query.",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "coffee-tea/beans"
            },
            "required": false,
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 60
            },
            "required": false,
            "name": "per_page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "relevance",
                "newest",
                "oldest",
                "price_asc",
                "price_desc",
                "rating"
              ],
              "description": "A whitelisted enum, never a raw sort expression. Defaults to relevance with a query and newest without one."
            },
            "required": false,
            "description": "A whitelisted enum, never a raw sort expression. Defaults to relevance with a query and newest without one.",
            "name": "sort",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "1500-4000",
              "description": "Minor units."
            },
            "required": false,
            "description": "Minor units.",
            "name": "price",
            "in": "query"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "required": false,
            "name": "in_stock",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of product cards with the generated facet panel.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ListingCard"
                      }
                    },
                    "page": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "perPage": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "degraded": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "page",
                        "perPage",
                        "total",
                        "totalPages",
                        "degraded"
                      ]
                    },
                    "facets": {
                      "type": [
                        "array",
                        "null"
                      ],
                      "items": {
                        "$ref": "#/components/schemas/Facet"
                      }
                    },
                    "ignoredFilters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "key",
                          "reason"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data",
                    "page",
                    "facets",
                    "ignoredFilters"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request was malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to this caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/products/{slug}": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "One product, with its attributes and variants.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The product.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to this caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/catalog/categories": {
      "post": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin catalogue"
        ],
        "summary": "Create a category.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "slug": {
                    "type": "string",
                    "maxLength": 80,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                  },
                  "parent": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^[0-9a-fA-F]{24}$",
                    "default": null
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "imagePublicId": {
                    "type": "string"
                  },
                  "order": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 0
                  },
                  "validationMode": {
                    "type": "string",
                    "enum": [
                      "lenient",
                      "strict"
                    ],
                    "default": "lenient"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "hidden"
                    ],
                    "default": "active"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "path": {
                          "type": "string",
                          "example": "coffee-tea/beans"
                        },
                        "parent": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ancestors": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "Root-first, including self."
                        },
                        "depth": {
                          "type": "integer"
                        },
                        "order": {
                          "type": "integer"
                        },
                        "description": {
                          "type": "string"
                        },
                        "imagePublicId": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active",
                            "hidden"
                          ]
                        },
                        "validationMode": {
                          "type": "string",
                          "enum": [
                            "lenient",
                            "strict"
                          ]
                        },
                        "attributeBindings": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "defId": {
                                "type": "string"
                              },
                              "key": {
                                "type": "string"
                              },
                              "required": {
                                "type": "boolean"
                              },
                              "order": {
                                "type": "integer"
                              },
                              "group": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "defId",
                              "key",
                              "required",
                              "order"
                            ]
                          }
                        },
                        "suppressedKeys": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "_id",
                        "name",
                        "slug",
                        "path",
                        "parent",
                        "ancestors",
                        "depth",
                        "order",
                        "status",
                        "validationMode",
                        "attributeBindings",
                        "suppressedKeys"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with something that already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin catalogue"
        ],
        "summary": "The whole category tree, hidden branches included.",
        "responses": {
          "200": {
            "description": "Root nodes, each with children.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminCategory"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}